+++ hvcc/generators/c2dpf/templates/DistrhoPluginInfo.h
@@ -3,46 +3,46 @@
 #pragma once
 
 #define DISTRHO_PLUGIN_NAME                 "{{name.replace('_', ' ')}}"
-{%- if meta.plugin_uri is defined %}
+{%- if meta.plugin_uri != None %}
 #define DISTRHO_PLUGIN_URI                  "{{meta.plugin_uri}}"
 {% else %}
 #define DISTRHO_PLUGIN_URI                  "urn:hvcc:{{name}}"
 {%- endif %}
-{%- if meta.plugin_clap_id is defined %}
+{%- if meta.plugin_clap_id != None %}
 #define DISTRHO_PLUGIN_CLAP_ID              "{{meta.plugin_clap_id}}"
 {% else %}
 #define DISTRHO_PLUGIN_CLAP_ID              "urn.hvcc.{{name}}"
 {%- endif %}
 #define DISTRHO_PLUGIN_NUM_INPUTS           {{num_input_channels}}
 #define DISTRHO_PLUGIN_NUM_OUTPUTS          {{num_output_channels}}
-#define DISTRHO_PLUGIN_IS_SYNTH             {{1 if num_output_channels > 0 and meta.midi_input is defined and meta.midi_input > 0 else 0}}
-#define DISTRHO_PLUGIN_HAS_UI               {{1 if meta.enable_ui is defined and meta.enable_ui is sameas true else 0}}
+#define DISTRHO_PLUGIN_IS_SYNTH             {{1 if num_output_channels > 0 and meta.midi_input > 0 else 0}}
+#define DISTRHO_PLUGIN_HAS_UI               {{1 if meta.enable_ui is sameas true else 0}}
 #define DISTRHO_PLUGIN_IS_RT_SAFE           1
 #define DISTRHO_PLUGIN_WANT_PROGRAMS        0
 #define DISTRHO_PLUGIN_WANT_STATE           0
 #define DISTRHO_PLUGIN_WANT_TIMEPOS         1
 #define DISTRHO_PLUGIN_WANT_FULL_STATE      0
-#define DISTRHO_PLUGIN_WANT_MIDI_INPUT      {{meta.midi_input if meta.midi_input is defined else 1}}
-#define DISTRHO_PLUGIN_WANT_MIDI_OUTPUT     {{meta.midi_output if meta.midi_output is defined else 1}}
-{%- if meta.lv2_info is defined %}
+#define DISTRHO_PLUGIN_WANT_MIDI_INPUT      {{1 if meta.midi_input is sameas true else 0}}
+#define DISTRHO_PLUGIN_WANT_MIDI_OUTPUT     {{1 if meta.midi_output is sameas true else 0}}
+{%- if meta.lv2_info != None %}
 #define DISTRHO_PLUGIN_LV2_CATEGORY         "{{meta.lv2_info}}"
 {%- endif %}
-{%- if meta.vst3_info is defined %}
+{%- if meta.vst3_info != None %}
 #define DISTRHO_PLUGIN_VST3_CATEGORIES      "{{meta.vst3_info}}"
 {%- endif %}
-{%- if meta.clap_info is defined %}
+{%- if meta.clap_info|length > 0 %}
 #define DISTRHO_PLUGIN_CLAP_FEATURES        "{{ meta.clap_info|join('", "') }}"
 {%- endif %}
 
 // for level monitoring
 #define DISTRHO_PLUGIN_WANT_DIRECT_ACCESS   0
 
-{%- if meta.enable_ui is defined %}
+{% if meta.enable_ui is sameas true %}
 // if you are using a UI you'll probably want to modify these settings to your needs
 #define DISTRHO_UI_USE_CUSTOM               1
 #define DISTRHO_UI_CUSTOM_INCLUDE_PATH      "DearImGui.hpp"
 #define DISTRHO_UI_CUSTOM_WIDGET_TYPE       DGL_NAMESPACE::ImGuiTopLevelWidget
-    {%- if meta.ui_size is defined %}
+    {%- if meta.ui_size != None %}
 #define DISTRHO_UI_DEFAULT_WIDTH            {{meta.ui_size.width}}
 #define DISTRHO_UI_DEFAULT_HEIGHT           {{meta.ui_size.height}}
     {% else %}
@@ -51,7 +51,7 @@
     {%- endif %}
 {%- endif %}
 
-{%- if meta.enable_modgui is defined %}
+{%- if meta.enable_modgui is sameas true %}
 #ifdef __MOD_DEVICES__
 #undef DISTRHO_PLUGIN_USES_MODGUI
 #define DISTRHO_PLUGIN_USES_MODGUI 1